Handling colors with BottomSheetHeader component
Changing colors is hard, so far, we had handled colors on
the javascript side, and that it was hard, and was not working very
well, we also have been using Animator to handle color transitions,
but Animator doesn't work with FloatingActionButton, because it's a
native component, also, there's not way to use native driven with
backgroundColor AFAIK.

But in the end, the code was a mess, and bloaded, too much set state,
and sometimes slow, the idea is to avoid react states and abstract it
on the native side.

I already tried to implement a similar approach before, but the main
problem was to connect all <Text> child components, which i gave up
and i moved to javascript and Animator instead.

Now i came with a better solution, BottomSheetHeader will automatically
find all ReactTextView instances (it includes <Text> and <Icon>) components
recursively and also FloatingActionButton and handles everything out the box.

And, the hole implementation lives directly on core of the behavior
with just a few lines of code, and it is much easier to handle states and
predict the target state before the animation itself.

There's still some stuff to be done, i'm trying to find a way to get
a style prop color (<Text style={{color: 'red'}} />) on the native side,
and so far i've only be able to get the color from `selectionColor`
prop that i stored to toggle colors back, i've described it here.
http://stackoverflow.com/questions/43061034/how-to-get-style-props-on-the-native-side-in-react-native