Merged AnchorSheetBehavior with BottomSheetBehavior
* This is possibly a breaking change, even though the things is working as
before, the biggest change here is that we are not using the support design
bottom sheet implementation anymore, instead, we are using our own
custom bottom sheet that is a fork but supports anchor state and color managment.

* Anchor state are now optional and disabled by default, it can be enabled
through `anchorEnabled` prop.

* BottomSheetBehaviorView inherits a NestedScrollView instead of a pure
RelativeLayout, this make things easier to work with our custom behavior
implementation, and others features such Backdrop and BottomSheetHeader,
but unfortunately, it breaks any ReactNestedScroll child, and i had to
change the ReactNestedScroll touch handlers and it should work as before.