Fixed non-translucents StatusBars
ScrollingAppBarLayout and MergedAppBarLayout now supports `translucent` prop.

The `translucent` prop just adjust the AppBar margins, and doesn't
change anything on the translucent itself, you should handle the
translucent directly on the react-native StatusBar component.

This also adds `barStyle` and `barStyleTransparent` props to set the
right `light-content` and `dark-content` the same as react-native.

Fixes #23